requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.6.4
azure-devops>=7.2.0
python-dateutil>=2.9.0.post0  # For better datetime handling 